fix: quest update
