+++ src/lib/battleABI.js
@@ -2,4 +2,4 @@
 import battleGameArtifact from "./BattleGameABI.json";
 
 export const battleGameABI = battleGameArtifact.abi;
-export const battleGameAddress = "0x7bf59FD7832FD14249a452Bc5969201A74934d8d";
+export const battleGameAddress = "0x557baa12c905C37bF76B94e53aE8904ec61AE5e3";
