+++ src/lib/creatureABI.js
@@ -2,4 +2,4 @@
 import CreatureNFT from "../../artifacts/contracts/CreatureNFT.sol/CreatureNFT.json";
 
 export const creatureABI = CreatureNFT.abi;
-export const creatureAddress = "0x74c1444D2Dc18433514883A39BBEda3C9815593f"; // 🔁 Replace if needed
+export const creatureAddress = "0x1390C5Bdcb64A2070144416f10Cc651474454AA0"; // 🔁 Replace if needed
